+++ tests/test_commands.py
@@ -337,11 +337,11 @@ class TreesappTester(unittest.TestCase):
         test_refpkg.slurp()
         self.assertIsInstance(test_refpkg.feature_annotations, dict)
         self.assertTrue('Function' in test_refpkg.feature_annotations)
-        self.assertEqual(5, len(test_refpkg.feature_annotations["Function"]))
+        self.assertEqual(6, len(test_refpkg.feature_annotations["Function"]))
         self.assertIsInstance(test_refpkg.feature_annotations["Function"][0], CladeAnnotation)
         for ca in test_refpkg.feature_annotations["Function"]:
             if ca.name == "Short-chain alkane oxidation":
-                self.assertEqual(2, len(ca.members))
+                self.assertEqual(5, len(ca.members))
 
         # Test view with a complex attribute (feature_annotations)
         view_command_list = ["view",