+++ treesapp/update_refpkg.py
@@ -329,9 +329,9 @@ def formulate_create_command(ts_updater: Updater, args, final_stage) -> list:
 def update_features(old_refpkg: ts_ref_pkg.ReferencePackage, new_refpkg: ts_ref_pkg.ReferencePackage) -> None:
     # Match leaf node identifiers to each other
     leaf_node_name_map = {}
-    old_leaf_desc_map = {leaf.description: leaf for leaf in old_refpkg.generate_tree_leaf_references_from_refpkg()}
+    old_leaf_desc_map = {leaf.description: leaf for leaf in old_refpkg._ref_leaves.values()}
     # Dictionary is indexed by the new leaf node names
-    for desc, leaf in {ln.description: ln for ln in new_refpkg.generate_tree_leaf_references_from_refpkg()}.items():
+    for desc, leaf in {ln.description: ln for ln in new_refpkg._ref_leaves.values()}.items():
         if desc in old_leaf_desc_map:
             leaf_node_name_map[old_leaf_desc_map[desc].number + '_' + old_refpkg.prefix] = leaf.number + '_' + new_refpkg.prefix
 
@@ -339,8 +339,12 @@ def update_features(old_refpkg: ts_ref_pkg.ReferencePackage, new_refpkg: ts_ref_
         new_refpkg.feature_annotations[feature] = []
         for clade_annot in old_refpkg.feature_annotations[feature]:  # type: clade_annotation.CladeAnnotation
             new_annotation = clade_annotation.CladeAnnotation(name=clade_annot.name, key=feature)
-            for leaf_node in clade_annot.members:
-                new_annotation.members.add(leaf_node_name_map[leaf_node])
+            # try:
+            for leaf_node, rank in clade_annot.members.items():
+                new_annotation.members[leaf_node_name_map[leaf_node]] = rank
+            # except AttributeError:
+            #     for leaf_node in clade_annot.members:
+            #         new_annotation.members[leaf_node_name_map[leaf_node]] = 7
             new_annotation.taxa = clade_annot.taxa
             new_annotation.colour = clade_annot.colour
             new_refpkg.feature_annotations[feature].append(new_annotation)